requirements.txt
google-api-python-client
discord.py
yt-dlp
python-dotenv
flask